+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/controller/v2/WorkflowV2ControllerTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.dolphinscheduler.api.controller;
+package org.apache.dolphinscheduler.api.controller.v2;
 
 import static org.apache.dolphinscheduler.common.constants.Constants.EMPTY_STRING;
